Improve speed of JSON format parser.
- parses floats always as doubles
- uses cached URIs for items and properties

+++ bundles/io.github.linkedfactory.kvin/src/test/java/io/github/linkedfactory/kvin/util/JsonFormatParserTest.java
@@ -48,7 +48,7 @@ public class JsonFormatParserTest {
             } else if (index == 3) {
                 assertTrue(t.value instanceof BigInteger);
             } else if (index == 4) {
-                assertTrue(t.value instanceof BigDecimal);
+                assertTrue(t.value instanceof Double);
             } else if (index == 5) {
                 assertTrue(t.value instanceof Long);
             } else if (index == 6) {
